frontend/account: call onAccountChanged after init/onStatusChanged
Before this change, onAccountChanged was called immediately when the
account was switched and the code changed, as it depended on the code,
which changed the callback, which led to it being called due to this line:

`useEffect(onAccountChanged, [onAccountChanged, status]);`

Calling this callback before `accountApi.init()` is incorrect and
leads to an error.

Instead of depending on the code and status, we pass it as params
instead.

We also await the result of `init()`, as technically it is also async
and could finis after `onAccountChanged()` was called.