frontend: improved portfolio chart on mobile
portfolio chart on mobile looked like it had uneven spacings
and whitespaces which made it not centered. This is due to
the vertical label (y-axis) drawn by the library.

This commit is inteded to fix that chart’s UI by partly re-styling
it for mobile, where the vertical label and side paddings are
removed.

To interact with the chart on mobile, long-press on the chart and
move around the chart. To stop interacting, users can simply lift
their finger (stop touching the chart).