backend/electrum: make Headers rpc call blocking/synchronous
The call-site already forced blocking anyway by use of the
channel. This propagates the error more directly and is easier to
handle.

The panics are turned into simple errors logs, as it's not critical -
the btc accounts can still function without the headers, just the
transactions would be missing their timestamp.