backend/btc/account: remove account lock, add fine grained locks
Locking the whole account is confusing, as it is not which struct
fields the lock covers and when to use it. Individual field locks are
much easier to reason about.

Functions that access .subaccounts (e.g. Info() and
GetUnusedReceiveAddresses()) are only valid after Initialize(). For
now as a shortcut, I simply return an empty value if not yet
initialized and have the frontend bail out, because it is hard or
impossible to achieve this state in the frontend, as one can only
click these buttons to access these endpoints after init.