backend/btc/account: remove potential HistoryStatus race conditions
The transaction history of an address is stored in the database, but
it's status (a hash derived from the history) was also stored in RAM
to decide if the address history has changed or not.

Being sure the DB value and the RAM value were in sync is not
trivial. Before, we used a mutex inside `onAddressStatus()`, but that
was not sufficient - the line

`address.HistoryStatus = history.Status()` could be set by another
goroutine before moving on to
`account.transactions.UpdateAddressHistory()`.

We can simplify this by only using the database as the source of
truth, avoiding all these synchronization issues.

The comment about the remaining race condition there could be solved
by using the same database transaction to check and update the status
instead of using two separate ones. Not done in this commit.

Currently the `getAddressHistory()` opens a writable DB transaction,
which acquires a write lock. The next commit changes this to use
read-locks only.

The TODOs for error handling are postponed to a comit which deals with
error handling in general, as there are many instances of unhandled
errors throughout the account handling.