backend/btc/electrum: use block-client-go library, remove jsonrpc
block-client-go is a new library that takes care of JSONRPC
connections, the Electrum protocol and failover managment. It was
written to replace the current electrum client and jsonrpc, which
suffers from many problems, all of which are fixed by block-client-go:

- saner API and error handling
- server.version protocol negotiation happens per server (currently it
  is applied for one server and used the same is used after failing
  over to another server, which is not correct)
- use of proper method timeouts (using `context`)
  separating out failover functionality, which previously was
  intermingled with the JSON RPC client
- properly unit tested (see unit tests in block-client-go)

The blockchain interface is simplified to remove a unused callback and
unused `err` param (it might have been useful before, but not with
block-client-go).