backend/moonpay: fix bug in widget language param
Moonpay widget takes a language param as input. Previously this was
passed as stored in `backend.UserLanguage`. The problem is that if
the set locale matches the system locale this param is empty, and the
language is not specified when calling the Moonpay widget.

Now if the userLanguage field is empty a language string retrieved by
the native locale of the system is used.

A function that extract only the first part of the locale has been added
in the new file `util/config/locale.go`.