frontend: pre-select region from user's native locale and config
to improve our UX, it would be nice to have a pre-selected value
for the region selector in the exchange / buy page.

This value is taken from user's config, which gets set the first time
a user selects a region. If it's not available yet, the value
(region) is taken from `native-locale` API.

If `native-locale` doesn't  provide a proper value, e.g
`es-es419`, then there will be no pre-selected / default
value for the selector.