frontend: fix flickering sidebar when navigating to device settings
Navigating to device settings triggered the sidebar to close, this
was the reason that it looked like the whole app window flickered.

Forcing the sidebar to close on will mount is not necessary as it
will open/close later depending on the device state.

Tested on medium screen (when hamburger is visible)
- reset device
- navigate through app
- open app with unlocked BB
- plugin uninitialized BB on waiting screen or app settings