backend/eth/account: fix race condition on tx-proposal
A race condition between the `Initialize()` and `TxProposal()` methods
was causing a deadlock were they were blocking each other in accessing
`account.Synchronizer` and `account.updateLock`. This recently caused
CI tests to timeout and fail from time to time.

This changes the behavior of `newTx()`, which is called by
`TxProposal()`, making it returning an errror instead of waiting when
the account is not yet synced. This should fix the race condition.