frontend: remove additional styling on buttons in exchange page
to match our app's styling and to maintain uniformity throughout
our app's UI, we should remove the additional/custom css applied
to the "back" and "next" button in the exchange page.

Also altered the buttons' container margin to match the
previous page's  button (the "next"  button) spacing, for the
same reason as the above.