frontend: add dark-mode toggle in app settings
Added a toggle to change darkmode in app settings.

When the user changes to a different mode than the OS it should
save it to config. If the user changes to the same mode as the OS
it should remove the config and use whatever the OS prefers.

This also adds a Darkmode component that is always mounted and in
charge of setting dark-mode CSS class to document body.

It also provides a getter function to keep track of which mode is
currently used, this does not use context or hook and cannot force
a re-render if for example the OS mode changes.

In the future this could be written as a hook, i.e. useDarkmode.