scripts: make dockerdev.sh and docker_exec.sh podman-compatible
podman is a (almost) drop-in replacment to docker that runs rootless
and without a daemon. Since it runs rootless, it maps the host user to
the root user inside the container. This clashes with the tricks used
for docker to mount the volume with the same host user id and group
id. This script removes these docker tricks when using podman.

golangci had to be updated as the previous version was not compatible
with 1.20.