frontend: rephrase sd card warning when creating a new wallet
There was no exlpanation of why the microSD card should be inserted
when creating a new wallet during the set wallet name step.

Moved step description next to the title followed by the warning
and added minimal information about what the SD Card will be used
for, i.e.  to create a backup of the wallet.