android: fix unwanted blocking of iframe links
Iframes integrated in the app have some external links that the user
could click to open in an external browser. Before this, the view url
were used to allow the links clicked in views inside the '/buy/ route.
For unknown reasons this recently stopped working (the retrieved url was
"about:blank"). Since it was not possible to retrive the view url in
other ways, this just creates a list of allowed domains that can be
pointed to.