frontend: refactor utxo component to functional component
Slightly changed how the component communicates with send.tsx,
it used a React ref that the parent (send) could clear the child.
Changed to just clear the selectedUTXOs map directly after send
or when the dialog opens (it will set the data again when it's
closed).

tested also that the UTXO list updated after sending which was
fixed in 582ad7e63e9b8d4ab5a700d4da90d938cfa75276