frontend: use conditional darkmode config in js
This is a small POC that gets darkmode info from darkmode.tsx
directly and changes chart layout to dark mode.

This has the drawback that the UI does not automatically re-render
when the user changes the OS settings as darkmode is just a variable
local to darkmode.tsx component. Once chart is a functional comp.
we can implement useDarkmode hook.