frontend: improve layout options of grid component
Column currently always adds margin-bottom but sometimes this is
not needed, i.e. if the children already add spacing.

Added support for className pro on Column component so that the
space can be manually added via CSS class.

Also added col prop to grid component to configure a single column.