backend/usb: only force single thread on macOS
In a35e009b we forced all USB communication to happen in a single
thread because macOS otherwise fails with obscure errors (see commit
msg).

For simplicity we applied this to all platforms. At least on Windows,
users ar experiencing issues where the BitBox02 would timeout during
password entry. It is not clear why, but since this was the only
change related to USB communication, we revert it for all platforms
except macOS.