frontend: fix matchmedia mock in tests
i18n/config.test had to mock matchmedia as at the time the dialogs
had an animation which supported reduce-motion mediaquery.
see cec2165d9e16eb074ee069c5272aca39c459c111

This was removed and since then the matchmedia.mock is not needed.

With useDarkmode (which is using useMediaquery internally) the Alert
component test now requires mocking window.matchMedia again, but
for some unknown reason the way that it is recommended to do the
mocking does not work, see recommended mock:
- https://jestjs.io/docs/manual-mocks#mocking-methods-which-are-not-implemented-in-jsdom

Removed old unused mock and added a version without defineProperty
to the alert test.