backend/accounts: introduce "hiddenBecauseUnused" accounts
In order to scan accounts until we find an empty account, we always
need to scan one account ahead of what is visible to the user.

For this, we introduce a flag "hiddenBecauseUnused", which is false by
default and only true if an account is added to be scanned in the
background. It is not shown to the user at all, until either:

- it is scanned and has a transaction history
- the user adds a new account in the settings - this activtes the
  hidden account if it was automatically added before.

Such hidden accounts are persisted for the same reason the regular
accounts are persisted: creating the entry requires the xpubs, which
requires a keystore (BitBox02), which is slow. By persisting it we do
not need to query the keystore for multiple xpubs everytime for the
hidden account, but just once.

The part where we add the hidden account automatically follows in the
next commit.