frontend: fix select dropdown colors on windows
Windows rendered the dropdown almost unreadable with light text on
white background. This is with user-defined darkmode but system
using light theme.

Reason: background color was hardcoded to white but font had no
color definition.

Changed to defining both background-color and text color.

This should have no effect on macOS as the dropdown are rendered
by the OS.