frontend/exchange: fix back button behavior
Back button on exchange selection page wasn't working correctly on
Android. Also the logic to choose where to land when pressing back was
based on the number of total accounts, which is wrong. Now it is based
on the number of total accounts with at least one available exchange.

Back button is hidden if there is only one supported account, since in
that case the account selection page is skipped.