frontend: fix regression api create-backup
createBackup was still using catch in case of success false, this
should have been part of 374bd5377b0021ab23f452657e1ddaf721524f63

This also fixes an issue when continuing the workflow after the
user aborted is-date-correct on the device. After canceling the
app showed confirm date on device but the device did not continue.